Refactoring - Created a function
